test (CelebrityFanazlyer): Add Reference Test File
This file will not ultimately be saved. It is only meant as a potential guide to accompany the tutorial: https://test-utils.vuejs.org/guide/

Modifications have been made so that it works with Quasar. As work progresses, the file is expected to change and then ultimately be deleted and be replaced by legitimate testing files.